Add toggle nav icon for mobile devices
On mobile devices Navs are hidden and are toggled by clicking an icon.
The show/hide mechanism of nav has been obtained using a checkbox. It
requires additional markup with no semantic meaning, but there's no
other way to implement it without add a javascript dependency.
The initial idea was to take advantage of the hover on the navbar, but
it has more problems than benefits, so I have agreed to the compromise
of the additional markup.
This is just a draft, styles and behaviors are not well defined yet.